+++ tests/Fixtures/HybridWidgetService.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Ngos\AdminCore\Tests\Fixtures;
+
+use Ngos\AdminCore\Services\CrudService;
+
+class HybridWidgetService extends CrudService
+{
+    public function __construct(HybridWidget $model)
+    {
+        $this->model = $model;
+    }
+}